# docs/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionTest": {
        "additionalProperties": false,
        "properties": {
          "auth_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when the source could not be reached.",
            "title": "Auth Ok"
          },
          "checked_at": {
            "title": "Checked At",
            "type": "string"
          },
          "families": {
            "items": {
              "$ref": "#/components/schemas/FamilyCapability"
            },
            "title": "Families",
            "type": "array"
          },
          "history_days": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Days of matching history found, up to 30.",
            "title": "History Days"
          },
          "kind": {
            "$ref": "#/components/schemas/SourceKind",
            "default": "prometheus"
          },
          "matched_series": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Prometheus: series currently matching all matchers.",
            "title": "Matched Series"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "reachable": {
            "title": "Reachable",
            "type": "boolean"
          }
        },
        "required": [
          "reachable",
          "auth_ok",
          "checked_at"
        ],
        "title": "ConnectionTest",
        "type": "object"
      },
      "ConnectionTestRequest": {
        "additionalProperties": false,
        "description": "Body of POST /api/projects/test-connection: a draft, possibly reusing stored secrets.",
        "properties": {
          "matchers": {
            "description": "Required with a Prometheus source.",
            "items": {
              "$ref": "#/components/schemas/LabelMatcher"
            },
            "maxItems": 10,
            "title": "Matchers",
            "type": "array"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reuse this project's stored secrets for omitted ones.",
            "title": "Project Id"
          },
          "source": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PrometheusSourceInput"
              },
              {
                "$ref": "#/components/schemas/CloudflareSourceInput"
              }
            ],
            "title": "Source"
          }
        },
        "required": [
          "source"
        ],
        "title": "ConnectionTestRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LabelMatcher": {
        "additionalProperties": false,
        "description": "Exact ``name=\"value\"`` matcher added to every selector of a project's queries.",
        "properties": {
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "value": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "name",
          "value"
        ],
        "title": "LabelMatcher",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "description": "RFC 9457 problem details (application/problem+json).",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
      "Project": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "credentials_readable": {
            "default": true,
            "description": "False when stored secrets cannot be decrypted (lost or changed key).",
            "title": "Credentials Readable",
            "type": "boolean"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "keep_reports": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Newest reports kept; null keeps all.",
            "title": "Keep Reports"
          },
          "matchers": {
            "items": {
              "$ref": "#/components/schemas/LabelMatcher"
            },
            "title": "Matchers",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "next_scheduled_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Next automatic analysis; null without a schedule.",
            "title": "Next Scheduled Run"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportSchedule"
              },
              {
                "type": "null"
              }
            ]
          },
          "sources": {
            "items": {
              "discriminator": {
                "mapping": {
                  "cloudflare": "#/components/schemas/CloudflareSource",
                  "prometheus": "#/components/schemas/PrometheusSource"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/PrometheusSource"
                },
                {
                  "$ref": "#/components/schemas/CloudflareSource"
                }
              ]
            },
            "title": "Sources",
            "type": "array"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "name",
          "matchers",
          "sources",
          "created_at",
          "updated_at"
        ],
        "title": "Project",
        "type": "object"
      },
      "ProjectInput": {
        "additionalProperties": false,
        "description": "Body of POST /api/projects and PUT /api/projects/{project_id}.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "keep_reports": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Keep only this many newest reports; older analyses are deleted automatically. Null keeps all.",
            "title": "Keep Reports"
          },
          "matchers": {
            "description": "Required (1-10) with a Prometheus source.",
            "items": {
              "$ref": "#/components/schemas/LabelMatcher"
            },
            "maxItems": 10,
            "title": "Matchers",
            "type": "array"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportSchedule"
              },
              {
                "type": "null"
              }
            ],
            "description": "Automatic analyses; null runs analyses on demand only."
          },
          "sources": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/PrometheusSourceInput"
                },
                {
                  "$ref": "#/components/schemas/CloudflareSourceInput"
                }
              ]
            },
            "maxItems": 2,
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectInput",
        "type": "object"
      },
      "ProjectList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ProjectList",
        "type": "object"
      },
      "ProjectSummary": {
        "additionalProperties": false,
        "description": "A project with its analysis activity (list and detail pages).",
        "properties": {
          "active_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisJob"
              },
              {
                "type": "null"
              }
            ],
            "description": "Queued or running job."
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "credentials_readable": {
            "default": true,
            "description": "False when stored secrets cannot be decrypted (lost or changed key).",
            "title": "Credentials Readable",
            "type": "boolean"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "keep_reports": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Newest reports kept; null keeps all.",
            "title": "Keep Reports"
          },
          "latest_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisJob"
              },
              {
                "type": "null"
              }
            ],
            "description": "Newest finished job."
          },
          "matchers": {
            "items": {
              "$ref": "#/components/schemas/LabelMatcher"
            },
            "title": "Matchers",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "next_scheduled_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Next automatic analysis; null without a schedule.",
            "title": "Next Scheduled Run"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "report_count": {
            "default": 0,
            "description": "Saved reports (deleted with it).",
            "minimum": 0.0,
            "title": "Report Count",
            "type": "integer"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportSchedule"
              },
              {
                "type": "null"
              }
            ]
          },
          "sources": {
            "items": {
              "discriminator": {
                "mapping": {
                  "cloudflare": "#/components/schemas/CloudflareSource",
                  "prometheus": "#/components/schemas/PrometheusSource"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/PrometheusSource"
                },
                {
                  "$ref": "#/components/schemas/CloudflareSource"
                }
              ]
            },
            "title": "Sources",
            "type": "array"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "name",
          "matchers",
          "sources",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectSummary",
        "type": "object"
      },
...
      "PrometheusSourceInput": {
        "additionalProperties": false,
        "properties": {
          "auth": {
            "discriminator": {
              "mapping": {
                "basic": "#/components/schemas/BasicAuthInput",
                "bearer": "#/components/schemas/BearerAuthInput",
                "none": "#/components/schemas/NoAuth"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/NoAuth"
              },
              {
                "$ref": "#/components/schemas/BearerAuthInput"
              },
              {
                "$ref": "#/components/schemas/BasicAuthInput"
              }
            ],
            "title": "Auth"
          },
          "kind": {
            "const": "prometheus",
            "default": "prometheus",
            "title": "Kind",
            "type": "string"
          },
          "tls_verify": {
            "default": true,
            "title": "Tls Verify",
            "type": "boolean"
          },
          "url": {
            "maxLength": 2048,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "PrometheusSourceInput",
        "type": "object"
      },
...
      "ReportSchedule": {
        "additionalProperties": false,
        "description": "Run an analysis automatically at ``time`` (local to ``timezone``) on ``weekdays``.\n\nEach run analyses the 24 h ending at the scheduled instant.",
        "properties": {
          "time": {
            "description": "Local time of day, HH:MM (24 h).",
            "examples": [
              "08:00"
            ],
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "title": "Time",
            "type": "string"
          },
          "timezone": {
            "default": "UTC",
            "description": "IANA time zone, e.g. Europe/Berlin.",
            "maxLength": 64,
            "title": "Timezone",
            "type": "string"
          },
          "weekdays": {
            "description": "Days to run on; all days by default.",
            "items": {
              "$ref": "#/components/schemas/Weekday"
            },
            "maxItems": 7,
            "minItems": 1,
            "title": "Weekdays",
            "type": "array"
          }
        },
        "required": [
          "time"
        ],
        "title": "ReportSchedule",
        "type": "object"
      },
...
      "SourceKind": {
        "enum": [
          "prometheus",
          "cloudflare"
        ],
        "title": "SourceKind",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "DevOps AI Assistant",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects": {
      "get": {
        "operationId": "list_projects_api_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project_api_projects_post",
        "parameters": [
          {
            "description": "Clone: reuse this project's stored secrets for omitted ones.",
            "in": "query",
            "name": "clone_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Clone: reuse this project's stored secrets for omitted ones.",
              "title": "Clone Of"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/test-connection": {
      "post": {
        "operationId": "test_connection_api_projects_test_connection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTest"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Test Connection",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{project_id}": {
      "delete": {
        "description": "Hard delete, including the project's analyses and reports.",
        "operationId": "delete_project_api_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "get_project_api_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "projects"
        ]
      },
      "put": {
        "operationId": "update_project_api_projects__project_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Update Project",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}